ct/l1: move l1 -> level_one
Review feedback: it's easier to read.
