+++ src/v/cloud_topics/level_one/state_update.h
@@ -11,7 +11,7 @@
 
 #include "absl/container/btree_set.h"
 #include "base/seastarx.h"
-#include "cloud_topics/l1/state.h"
+#include "cloud_topics/level_one/state.h"
 #include "container/chunked_hash_map.h"
 #include "container/fragmented_vector.h"
 #include "model/fundamental.h"
